requirements.txt
pandas==1.5.3
requests==2.28.2
psycopg2-binary==2.9.6
apache-airflow==2.8.4
apache-airflow-providers-smtp==1.1.0
